+++ packages/components/buttons/icon-button/index.js
@@ -1,3 +1,3 @@
-export { default } from '../../../../src/components/buttons/icon-button';
+export { default } from './src';
 
 export { default as version } from './version';
