+++ packages/components/buttons/icon-button/src/icon-button.story.js
@@ -3,9 +3,9 @@ import { MemoryRouter, Link } from 'react-router-dom';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, select, text } from '@storybook/addon-knobs/react';
-import Section from '../../../../.storybook/decorators/section';
-import * as icons from '../../icons';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import * as icons from '../../../../../src/components/icons';
+import Readme from '../README.md';
 import IconButton from './icon-button';
 
 const iconNames = Object.keys(icons);
