+++ packages/components/buttons/icon-button/src/icon-button.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { IconButton, InformationIcon } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 export const routePath = '/icon-button';
